Replace DWORD with WORD.
WORD is 16 bit, DWORD is 32 bit.

+++ src/inc/pendual.h
@@ -34,7 +34,7 @@
 #include <stdint.h>
 #else
 typedef BYTE uint8_t;
-typedef DWORD uint16_t;
+typedef WORD uint16_t;
 typedef UINT32 uint32_t;
 #endif
 #include <stdlib.h>
