+++ unittest.c
@@ -32,7 +32,7 @@
 #endif
 #define inline __inline
 typedef BYTE uint8_t;
-typedef DWORD uint16_t;
+typedef WORD uint16_t;
 typedef UINT32 uint32_t;
 #endif
